Introduce SceneFinder and rename TranscriptionFinder
Rename TranscriptionFinder to TextSegmentsFinder and update imports/usages across handlers and tests; add a new SceneFinder to centralize scene-cut queries and move scene-fetching logic out of SceneSnapService. Refactor SceneSnapService and handlers to call SceneFinder.fetch_scene_cuts. Simplify AdjustVideoClipHandler: use SegmentWithTimes types, return concrete segment structures, and split segment lookup into helper methods. Minor response formatting change for transcription output and add EXISTS key to ElasticsearchQueryKeys.
