+++ bot/services/scene_snap/scene_snap_service.py
@@ -1,21 +1,17 @@
 import bisect
 import logging
 from typing import (
-    Any,
-    Dict,
     List,
     Tuple,
     Union,
 )
 
-from bot.search.elastic_search_manager import ElasticSearchManager
+from bot.search.scene_finder import SceneFinder
 from bot.types import (
     ClipSegment,
     ElasticsearchSegment,
 )
 from bot.utils.constants import (
-    ElasticsearchKeys,
-    ElasticsearchQueryKeys,
     EpisodeMetadataKeys,
     SegmentKeys,
 )
@@ -23,12 +19,6 @@ from bot.utils.log import log_system_message
 
 
 class SceneSnapService:
-    __SCENE_INFO_FIELD = "scene_info"
-    __SCENE_NUMBER_FIELD = "scene_info.scene_number"
-    __SCENE_START_TIME = "scene_info.scene_start_time"
-    __SCENE_END_TIME = "scene_info.scene_end_time"
-    __UNIQUE_SCENES_AGG = "unique_scenes"
-    __SCENE_DATA_AGG = "scene_data"
     __KEYFRAME_INTERVAL = 0.5
 
     @staticmethod
@@ -36,91 +26,6 @@ class SceneSnapService:
         if is_start:
             return boundary + SceneSnapService.__KEYFRAME_INTERVAL
         return boundary - SceneSnapService.__KEYFRAME_INTERVAL
-
-    @staticmethod
-    def __build_scene_cuts_query(season: int, episode_number: int) -> dict:
-        return {
-            ElasticsearchQueryKeys.SIZE: 0,
-            ElasticsearchQueryKeys.QUERY: {
-                ElasticsearchQueryKeys.BOOL: {
-                    ElasticsearchQueryKeys.MUST: [
-                        {
-                            ElasticsearchQueryKeys.TERM: {
-                                f"{EpisodeMetadataKeys.EPISODE_METADATA}.{EpisodeMetadataKeys.SEASON}": season,
-                            },
-                        },
-                        {
-                            ElasticsearchQueryKeys.TERM: {
-                                f"{EpisodeMetadataKeys.EPISODE_METADATA}.{EpisodeMetadataKeys.EPISODE_NUMBER}": episode_number,
-                            },
-                        },
-                        {"exists": {"field": SceneSnapService.__SCENE_INFO_FIELD}},
-                    ],
-                },
-            },
-            ElasticsearchQueryKeys.AGGS: {
-                SceneSnapService.__UNIQUE_SCENES_AGG: {
-                    ElasticsearchQueryKeys.TERMS: {
-                        ElasticsearchQueryKeys.FIELD: SceneSnapService.__SCENE_NUMBER_FIELD,
-                        ElasticsearchQueryKeys.SIZE: 2000,
-                    },
-                    ElasticsearchQueryKeys.AGGS: {
-                        SceneSnapService.__SCENE_DATA_AGG: {
-                            ElasticsearchQueryKeys.TOP_HITS: {
-                                ElasticsearchQueryKeys.SIZE: 1,
-                                ElasticsearchQueryKeys.SOURCE: {
-                                    ElasticsearchQueryKeys.INCLUDES: [
-                                        SceneSnapService.__SCENE_START_TIME,
-                                        SceneSnapService.__SCENE_END_TIME,
-                                    ],
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        }
-
-    @staticmethod
-    def __extract_cuts_from_buckets(buckets: List[Dict[str, Any]]) -> List[float]:
-        raw_cuts = []
-        for bucket in buckets:
-            hits = bucket[SceneSnapService.__SCENE_DATA_AGG][ElasticsearchKeys.HITS][ElasticsearchKeys.HITS]
-            if not hits:
-                continue
-            scene_info = hits[0][ElasticsearchKeys.SOURCE].get(SceneSnapService.__SCENE_INFO_FIELD, {})
-            start = scene_info.get("scene_start_time")
-            end = scene_info.get("scene_end_time")
-            if start is not None:
-                raw_cuts.append(float(start))
-            if end is not None:
-                raw_cuts.append(float(end))
-        return raw_cuts
-
-    @staticmethod
-    async def fetch_scene_cuts(
-        series_name: str,
-        season: int,
-        episode_number: int,
-        logger: logging.Logger,
-    ) -> List[float]:
-        try:
-            es = await ElasticSearchManager.connect_to_elasticsearch(logger)
-            index = f"{series_name}_text_segments"
-            query = SceneSnapService.__build_scene_cuts_query(season, episode_number)
-            response = await es.search(index=index, body=query)
-            buckets = response[ElasticsearchKeys.AGGREGATIONS][SceneSnapService.__UNIQUE_SCENES_AGG][ElasticsearchKeys.BUCKETS]
-            raw_cuts = SceneSnapService.__extract_cuts_from_buckets(buckets)
-            scene_cuts = sorted(set(raw_cuts))
-            await log_system_message(
-                logging.INFO,
-                f"Fetched {len(scene_cuts)} scene cuts for S{season:02d}E{episode_number:02d} in '{series_name}'",
-                logger,
-            )
-            return scene_cuts
-        except Exception as e:
-            await log_system_message(logging.WARNING, f"Failed to fetch scene cuts: {e}", logger)
-            return []
 
     @staticmethod
     def snap_boundaries(
@@ -166,7 +71,7 @@ class SceneSnapService:
                 await log_system_message(logging.WARNING, "Missing episode metadata for scene snap", logger)
                 return clip_start, clip_end
 
-            scene_cuts = await SceneSnapService.fetch_scene_cuts(series_name, season, episode_number, logger)
+            scene_cuts = await SceneFinder.fetch_scene_cuts(series_name, season, episode_number, logger)
 
             if not scene_cuts:
                 return clip_start, clip_end
